Move UniformInt to a new submodule
